+++ src/views/Contractor.vue
@@ -10,49 +10,51 @@
       <v-row>
         <v-col>
           <h3>Полное название организации</h3>
-          <v-text-field v-model="contractor.account_name" placeholder="Введите полное название организации" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.account_name" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Юридический адрес</h3>
-          <v-text-field v-model="contractor.requisites.legal_address" placeholder="Введите юридический адрес" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.legal_address" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Почтовый адрес</h3>
-          <v-text-field v-model="contractor.requisites.real_address" placeholder="Введите почтовый адрес" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.real_address" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>ФИО руководителя организации</h3>
-          <v-text-field v-model="contractor.requisites.full_name" placeholder="Введите ФИО руководителя организации" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.full_name" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Должность руководителя</h3>
-          <v-text-field v-model="contractor.requisites.director_role_name" placeholder="Введите должность руководителя" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.director_role_name" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Телефон контактный</h3>
-          <v-text-field v-model="contractor.requisites.contact_phone" placeholder="Введите телефон контактный" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.contact_phone" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Email руководителя</h3>
-          <v-text-field v-model="contractor.requisites.director_email" placeholder="Введите email руководителя" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.director_email" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>ОГРН</h3>
-          <v-text-field v-model="contractor.requisites.ogrn" placeholder="Введите ОГРН" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.ogrn" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>ИНН</h3>
-          <v-text-field v-model="contractor.requisites.inn" placeholder="Введите ИНН" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.inn" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>КПП</h3>
-          <v-text-field v-model="contractor.requisites.kpp" placeholder="Введите КПП" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.kpp" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>БИК</h3>
-          <v-text-field v-model="contractor.requisites.bik" placeholder="Введите БИК" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.bik" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Наименование банка</h3>
-          <v-text-field v-model="contractor.requisites.bank_name" placeholder="Введите наименование банка" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.bank_name" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Расчетный счет</h3>
-          <v-text-field v-model="contractor.requisites.billing_account" placeholder="Введите расчетный счет" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.billing_account" placeholder="не указан" :disabled="!editable"></v-text-field>
 
           <h3>Корреспондентский счет</h3>
-          <v-text-field v-model="contractor.requisites.bank_account" placeholder="Введите корреспондентский счет" :disabled="!editable"></v-text-field>
+          <v-text-field v-model="contractor.requisites.bank_account" placeholder="не указан" :disabled="!editable"></v-text-field>
 
-          <v-btn depressed color="success" @click="updateContractor" class="mr-3">Обновить</v-btn>
-          <v-btn depressed color="error" @click="deleteContractor" class="mr-3">Удалить</v-btn>
+          <div v-if="editable">
+            <v-btn depressed color="success" @click="updateContractor" class="mr-3">Обновить</v-btn>
+            <v-btn depressed color="error" @click="deleteContractor" class="mr-3">Удалить</v-btn>
+          </div>
 
         </v-col>
 
@@ -79,6 +81,7 @@
             icon
             outlined
             rounded
+            @click="print()"
             class="ml-4">
             <v-icon>mdi-printer-outline</v-icon>
           </v-btn>
@@ -105,7 +108,10 @@ export default {
       this.$store.dispatch('updateContractor')
     },
     deleteContractor() {
-      this.$store.dispatch('deleteContractor')
+      confirm('Вы уверены что хотите удалить контрагента? Вернуть его уже будет нельзя!') && this.$store.dispatch('deleteContractor')
+    },
+    print() {
+      window.print()
     }
   },
   computed: {
@@ -115,6 +121,9 @@ export default {
   },
   created () {
     this.$store.dispatch('getContractor', this.id)
+  },
+  beforeDestroy () {
+    this.$store.commit('setContractor', {})
   }
 }
 </script>
